pyproject.toml
[tool.poetry]
name = "drf-learning-api"
version = "0.1.0"
description = ""
authors = ["RafaelEmery <rafael.emerycade@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.7"
django = "2.2.9"
djangorestframework = "^3.12.4"
django-filter = "^2.4.0"
markdown = "<3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
